LPS-65012 Fix internal package in wiki-engine-jspwiki
