+++ modules/apps/wiki-engine-jspwiki/wiki-engine-jspwiki/src/main/java/com/liferay/wiki/engine/jspwiki/internal/LiferayURLConstructor.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.wiki.engine.jspwiki;
+package com.liferay.wiki.engine.jspwiki.internal;
 
 import com.ecyrd.jspwiki.WikiContext;
 import com.ecyrd.jspwiki.url.URLConstructor;
