+++ modules/apps/wiki-engine-jspwiki/wiki-engine-jspwiki/src/main/resources/com/liferay/wiki/engine/jspwiki/dependencies/jspwiki.properties
@@ -9,7 +9,7 @@ creole.imagePlugin.para.XXL=width='1024'
 creole.renderFreeExternalLinks=true
 jspwiki.aclManager=com.ecyrd.jspwiki.auth.acl.DefaultAclManager
 jspwiki.applicationName=LiferayWiki
-jspwiki.attachmentProvider=com.liferay.wiki.engine.jspwiki.LiferayAttachmentProvider
+jspwiki.attachmentProvider=com.liferay.wiki.engine.jspwiki.internal.LiferayAttachmentProvider
 jspwiki.authorizer=com.ecyrd.jspwiki.auth.authorize.GroupManager
 jspwiki.basicAttachmentProvider.storageDir=/tmp
 jspwiki.breakTitleWithSpaces=false
@@ -21,7 +21,7 @@ jspwiki.defaultprefs.template.timezone=GMT+01:00
 jspwiki.diffProvider=TraditionalDiffProvider
 jspwiki.encoding=UTF-8
 jspwiki.fileSystemProvider.pageDir=/tmp/
-jspwiki.groupdatabase=com.liferay.wiki.engine.jspwiki.LiferayGroupDatabase
+jspwiki.groupdatabase=com.liferay.wiki.engine.jspwiki.internal.LiferayGroupDatabase
 jspwiki.groupdatabase.created=created
 jspwiki.groupdatabase.creator=creator
 jspwiki.groupdatabase.datasource=jdbc/GroupDatabase
@@ -46,7 +46,7 @@ jspwiki.interWikiRef.WikiWikiWeb=http://c2.com/cgi/wiki?%s
 jspwiki.interWikiRef.Wikipedia=http://www.wikipedia.com/wiki/%s
 jspwiki.lockExpiryTime=60
 jspwiki.mail.jndiname=mail/MailSession
-jspwiki.pageProvider=com.liferay.wiki.engine.jspwiki.LiferayPageProvider
+jspwiki.pageProvider=com.liferay.wiki.engine.jspwiki.internal.LiferayPageProvider
 jspwiki.plugin.searchPath=com.liferay.wiki.engine.impl.jspwiki.plugin
 jspwiki.renderingManager.useCache=false
 jspwiki.rss.channelDescription=Oh poor me, my owner has not set a channel description at all. Pity me.
@@ -54,7 +54,7 @@ jspwiki.rss.channelLanguage=en-us
 jspwiki.rss.fileName=rss.rdf
 jspwiki.rss.generate=false
 jspwiki.rss.interval=3600
-jspwiki.searchProvider=com.liferay.wiki.engine.jspwiki.LiferaySearchProvider
+jspwiki.searchProvider=com.liferay.wiki.engine.jspwiki.internal.LiferaySearchProvider
 jspwiki.security=none
 jspwiki.specialPage.CreateGroup=NewGroup.jsp
 jspwiki.specialPage.FindPage=Search.jsp
@@ -68,7 +68,7 @@ jspwiki.translatorReader.inlinePattern.1=no-image-inlined
 jspwiki.translatorReader.matchEnglishPlurals=true
 jspwiki.translatorReader.useAttachmentImage=false
 jspwiki.translatorReader.useOutlinkImage=false
-jspwiki.urlConstructor=com.liferay.wiki.engine.jspwiki.LiferayURLConstructor
+jspwiki.urlConstructor=com.liferay.wiki.engine.jspwiki.internal.LiferayURLConstructor
 jspwiki.usePageCache=false
 jspwiki.userdatabase=com.ecyrd.jspwiki.auth.user.XMLUserDatabase
 jspwiki.userdatabase.created=created
